experiments/tts-api-test/requirements.txt
elevenlabs>=1.9.0
python-dotenv>=1.0.1
